Added clearLoggedInIdentifiers to AppAnalytics and moved AppAnalytics into analytics module
