+++ core/analytics/src/main/java/org/mozilla/social/core/analytics/AppAnalytics.kt
@@ -1,4 +1,4 @@
-package org.mozilla.social
+package org.mozilla.social.core.analytics
 
 import android.content.Context
 import org.mozilla.social.core.analytics.Analytics
@@ -11,4 +11,6 @@ class AppAnalytics(private val analytics: Analytics) {
     fun appBackgrounded() = analytics.appBackgrounded()
 
     fun initialize(context: Context) = analytics.initialize(context)
+
+    fun clearLoggedInIdentifiers() = analytics.clearLoggedInIdentifiers()
 }
